pkg wip: version that uses tags for rooting submodules.
This version feels by far the most natural and least hacky of the
approaches I've tried so far. I think this is the way to go. Note
that you don't get *that* many tags since it's only one tag per
package version, not a tag per checkpoint or even a tag per unique
set of submodule versions.