load, require: make ".jl" suffix optional.
This is another step on the road to making module names correspond
to the name you require, as in:

    require("Foo")
    using Foo

Since packages explicitly declare what they require, we may be able
to just omit require statements altogether and let them be implied
but each packages declarative REQUIRE files, which also contain
version information. One wrinkle in that idea is that some packages
may dynamically decide at runtime what to require or not. Not sure
how to deal with that possibility, but we can worry about it later.