Pkg.build: this is the long-awaited Pkg.fixup replacement.
Fundamentally, it's not really that different. Instead of trying to
topologically sort packages, I just install them in alphabetical
order and then use depth-first recursion in build to make sure that
dependencies are built before the things that depend on them. If
two packages depend on each other circularly, then one of them will
get build firts, which may not work; we may have to deal with that
at some point, but let's worry about that if and when it happens.

Philosophically, this change takes the approach that as far as the
package installation machinery is concerned, build failures aren't
really it's problem – it will run the build script for a package
when that package is installed or updated, but if the build script
fails, the package is still considered to be installed. A big, red
warning is printed for the user so that they know the package is
not likely to work, but the package is left installed so that they
can go into its directory and try to fix it. This approach seems
likely to make installing packages with binary dependencies a lot
less annoying.