Pkg.success: use awful `git update-index --really-refresh` again.
This was horribly broken; apparently the fact that autorefreshindex
doesn't apply to plumbing commands means that they *default* to the
broken behavior. So now we use the plumbing commands and force the
index to be up-to-date by doing `git update-index --really-refresh`
before every Git.success command. Close #4349. Kill it with fire.

. I give up. This is a completely absurd
way to do this