Support git-lfs (large file storage) (#64)
* Support git-lfs (large file storage)

* Update readme to state git LFS support

* Silence lfs git tests

* Check for CI private key in test

* fix check on CI private key

* update env var check

* Read path from `git_cmd`, not global

The path can be modified by preceding code, so we need to read it from `git_cmd`, append the `git-lfs` directory, then write back to `git_cmd`.

* Try using `windows_verbatim` if git-lfs loaded

CI issue could be related to single quotes in `cmd.exec` entries?

* try removing single quotes from cmd.exec

* revert quote filtering scheme

* bump Julia compat to 1.8

Need bugfix to multiple calls to `addenv` on Windows

* julia compat revision

* Add CI job for Windows with Julia = 1.7.3

the `version: 'min'` jobs hit Julia 1.6.6 but we should also test against 1.7.3 on Windows to ensure the `addenv` calls are working

* Implement Mose's suggestions

Co-authored-by: Mosè Giordano <765740+giordano@users.noreply.github.com>